fix: :bug: 파라미터 Long -> Integer로 변경
1. 리소스 관점에서 Long 타입보다는 Integer 타입이 좋을것으로 보여 변경

+++ app/shop/src/main/java/com/shoppingmall/api/CartRestController.java
@@ -81,7 +81,7 @@ public class CartRestController {
     }
 
     @PutMapping("/cart/{cartId}")
-    public ResponseEntity<CommonResponse> deleteCartItem(@PathVariable("cartId") Long cartId) {
+    public ResponseEntity<CommonResponse> deleteCartItem(@PathVariable("cartId") Integer cartId) {
         Member member = SecurityUtils.getCurrentMember()
                 .orElseThrow(() -> new AccessDeniedException("인증된 사용자 정보를 찾을 수 없습니다."));
