+++ app/shop/src/main/java/com/shoppingmall/service/ProductService.java
@@ -56,14 +56,14 @@ public class ProductService {
         return new PagingResponseDto<>(products, pagination);
     }
 
-    public ProductDetailResponseDto getProductByProductId(Long productId) {
+    public ProductDetailResponseDto getProductByProductId(Integer productId) {
         return productMapper.getProductByProductId(productId)
                 .map(ProductDetailResponseDto::toDto)
                 .orElse(new ProductDetailResponseDto());
     }
 
     @Transactional
-    public Long saveProducts(ProductSaveRequestDto productRequestDto) {
+    public int saveProducts(ProductSaveRequestDto productRequestDto) {
         Product product = productRequestDto.toEntity();
         int responseCode = productMapper.saveProducts(product);
         if (responseCode == 0) {
@@ -90,7 +90,7 @@ public class ProductService {
         return product.getProductId();
     }
 
-    public int saveFiles(Long productId, List<FileSaveRequestDto> files) {
+    public int saveFiles(Integer productId, List<FileSaveRequestDto> files) {
         if (CollectionUtils.isEmpty(files) || files.get(0) == null || productId == null) {
             return 0;
         }
@@ -121,7 +121,7 @@ public class ProductService {
         }
     }
 
-    private int updateFilesByProductId(Long productId, List<MultipartFile> files) {
+    private int updateFilesByProductId(Integer productId, List<MultipartFile> files) {
         if (productFileMapper.countProductFileByProductId(productId) > 0) {
             int responseCode = productFileMapper.deleteFilesByProductId(productId); // [HINT] DB에 존재하는 파일 정보 삭제 (SOFT DELETE)
             if (responseCode > 0) {
@@ -138,7 +138,7 @@ public class ProductService {
     }
 
     @Transactional
-    public void deleteProduct(Long productId) {
+    public void deleteProduct(Integer productId) {
         int responseCode = productMapper.deleteProduct(productId);
 
         // 상품 삭제 성공하지 않았을 경우, 추가 작업 중단 후 메서드 종료
@@ -158,7 +158,7 @@ public class ProductService {
         }
     }
 
-    private List<FileResponseDto> getFileResponseDtos(long productId) {
+    private List<FileResponseDto> getFileResponseDtos(Integer productId) {
         return productFileMapper.getFilesByProductId(productId)
                 .stream()
                 .map(ProductFileResponseDto::toDto)
